feat(course,lecture): add CRUD endpoints for course and lecture management
- Add Course and Lecture REST APIs
- Implement create, read, update, delete operations
- Add service and repository layers
- Add request/response DTOs
- Add validation and exception handling

+++ backend/src/main/java/com/lmsrag/backend/repository/UserRepository.java
@@ -5,7 +5,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+/**
+ * Repository cho entity User.
+ * Sửa từ JpaRepository<User, Integer> thành Long để khớp với @Id Long id trong entity.
+ */
+public interface UserRepository extends JpaRepository<User, Long> {
     Optional<User> findByEmail(String email);
     boolean existsByEmail(String email);
 }
